refactor: remove KleinanzeigenBot root re-export
